mpt_extension_sdk/pyproject.toml
[project]
name = "mpt-extension-sdk"
version = "6.0.1"
description = "MPT Extension SDK v6"
authors = [
    { name = "SoftwareOne AG" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]==0.131.*",
  "mpt-api-client==5.4.*",
  "mrok==0.9.6",
  "opentelemetry-api==1.40.*",
  "opentelemetry-exporter-otlp-proto-http==1.40.*",
  "opentelemetry-instrumentation-fastapi==0.61b0",
  "opentelemetry-instrumentation-httpx==0.61b0",
  "opentelemetry-instrumentation-logging==0.61b0",
  "opentelemetry-sdk==1.40.*",
  "pydantic==2.12.*",
  "typer==0.24.*",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]

[project.optional-dependencies]
azure-monitor = [
  "azure-monitor-opentelemetry-exporter==1.0.0b51",
]

[project.scripts]
swoext = "mpt_extension_sdk.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["mpt_extension_sdk"]

[tool.hatch.build.targets.wheel]
packages = ["mpt_extension_sdk"]
